separate jo models
