+++ src/flaskdoc/pallets/app.py
@@ -7,7 +7,6 @@ import pkg_resources
 import yaml
 
 from flaskdoc import swagger
-from flaskdoc.jo.schema import schema_factory
 from flaskdoc.pallets import plugins
 from flaskdoc.pallets.blueprints import Blueprint
 from flaskdoc.pallets.mixin import SwaggerMixin
@@ -157,7 +156,7 @@ def get_api_docs(app):
             pi = parse_specs(rule, spec, api)
             pi.description = docs
             api.paths.add(plugins.parse_flask_rule(rule.rule), pi)
-    api.components["schemas"] = schema_factory.components
+    api.components["schemas"] = swagger.SCHEMA_FACTORY.components
     return 1
